♻️ consolidate remaining use_cases imports and lift xfail
